requirements.txt
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
idna==3.11
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
